# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Diliké",
        "version": "1.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "liste des categories",
                "description": "liste de de toutes les categories",
                "operationId": "categories_list",
                "responses": {
                    "200": {
                        "description": "Afficher la liste de de toutes les categories"
                    }
                }
            },
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Créer une categorie",
                "description": "Inscription d'une categorie",
                "operationId": "categories",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Plat Attié"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Obtenir une seul categories",
                "description": "Obtenir une seul categories",
                "operationId": "one_categorie",
                "responses": {
                    "200": {
                        "description": "Obtenir une seul categories"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Mise à jour d'une categorie",
                "description": "Mise à jour d'une categorie",
                "operationId": "categories_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Plat Attié"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Supprimer une categorie",
                "description": "Suprrime une categorie",
                "operationId": "delete_categories",
                "responses": {
                    "200": {
                        "description": "Supprime une categorie"
                    }
                }
            }
        },
...
    }
}